Allow staging project for file distribution
