+++ docs/file_distribution/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
-## Unreleased
+## 1.0.1
+
+* Updates to support `staging_project_id` config
 * Updates to Python 3.12 and switches to use `fw-gear` instead of `flywheel-gear-toolkit` (now deprecated)
 
 ## 1.0.0
